requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
contourpy==1.2.1
cramjam==2.8.3
cycler==0.12.1
fonttools==4.53.1
fsspec==2024.6.1
graphviz==0.20.3
idna==3.7
kiwisolver==1.4.5
matplotlib==3.9.1
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
py4j==0.10.9.7
pyparsing==3.1.2
pyspark==3.5.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
tqdm==4.66.4
tzdata==2024.1
urllib3==2.2.2
